Final Keylogger embedded Image
CAUTION: THIS FILE CONTAINS A KEY LOGGER
Disclaimer: This project contains materials which are meant for educational purpose. Under no circumstance shall I be Liable for any indirenct, incidental, consequential, special or exemplary damages arising out of or in connection with your access or use of or inability to access or use the application or any contents of the repository.